move montagu tests to custom config tests
