+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/app_start/ErrorHandler.kt
@@ -69,6 +69,7 @@ class ErrorHandler(freemarkerEngine: FreeMarkerEngine)
     {
         logger.warn("For request ${req.uri()}, a ${error::class.simpleName} occurred with the following problems: ${error.problems}")
 
+        val path = req.pathInfo()
         if (req.pathInfo().startsWith(Router.apiUrlBase))
         {
             apiErrorHandler.handleError(error, res)
