+++ src/customConfigTests/src/test/kotlin/org/vaccineimpact/orderlyweb/customConfigTests/RequestHelper.kt
@@ -1,13 +1,18 @@
 package org.vaccineimpact.orderlyweb.customConfigTests
 
+import com.beust.klaxon.JsonObject
+import com.beust.klaxon.Parser
 import khttp.responses.Response
+import khttp.structures.authorization.BasicAuthorization
 import org.vaccineimpact.orderlyweb.ContentTypes
 import org.vaccineimpact.orderlyweb.db.AppConfig
 import org.vaccineimpact.orderlyweb.security.WebTokenHelper
+import org.vaccineimpact.orderlyweb.security.clients.MontaguIndirectClient
 
 class RequestHelper
 {
     val apiBaseUrl: String = "http://localhost:${AppConfig()["app.port"]}/api/v1"
+    val webBaseUrl: String = "http://localhost:${AppConfig()["app.port"]}"
 
     fun get(url: String, contentType: String = ContentTypes.json,
             userEmail: String): Response
@@ -15,6 +20,31 @@ class RequestHelper
         val token = generateToken(userEmail)
         val headers = standardHeaders(contentType).withAuthorizationHeader(token)
         return get(apiBaseUrl + url, headers)
+    }
+
+    fun loginWithMontagu(): JsonObject
+    {
+        // these user login details are set up in ./dev/run-dependencies.sh
+        val auth = BasicAuthorization("test.user@example.com", "password")
+        val response = khttp.post("${AppConfig()["montagu.api_url"]}/authenticate/",
+                data = mapOf("grant_type" to "client_credentials"),
+                auth = auth
+        )
+        val text = response.text
+        println(text)
+        return Parser().parse(StringBuilder(text)) as JsonObject
+    }
+
+    fun getWithMontaguCookie(
+            url: String,
+            contentType: String = "text/html"
+    ): Response
+    {
+        val token = loginWithMontagu()["access_token"]
+        val cookieName = MontaguIndirectClient.cookie
+        val headers = standardHeaders(contentType) +
+                mapOf("Cookie" to "$cookieName=$token")
+        return khttp.get(webBaseUrl + url, headers, allowRedirects = false)
     }
 
     private fun standardHeaders(contentType: String): Map<String, String>
