+++ src/customConfigTests/src/test/resources/config.properties
@@ -16,6 +16,7 @@ app.port=${app_port}
 app.name=${app_name}
 app.email=${app_email}
 
+montagu.url=${montagu_url}
 montagu.api_url=${montagu_api_url}
 
 auth.provider=montagu
